--- src/lib/util/functions/interactionErrorHandler.ts
@@ -1,28 +1,0 @@
-import { container, Events, InteractionHandlerError, InteractionHandlerParseError, UserError } from '@sapphire/framework';
-import { DiscordAPIError, HTTPError } from 'discord.js';
-
-import { errorEmbedUser, getWarnError, handleUserError, IgnoredCodes, sendErrorToUser } from './errorHelper';
-
-export function interactionErrorHandler(error: Error, { interaction }: InteractionHandlerError | InteractionHandlerParseError) {
-	if (typeof error === 'string') return sendErrorToUser(interaction, errorEmbedUser(error));
-	if (error instanceof UserError) return handleUserError(interaction, error);
-
-	const { client, logger } = container;
-
-	if (error.name === 'AbortError' || error.message === 'Internal Server Error') {
-		logger.warn(`${getWarnError(interaction)} (${interaction.user.id}) | ${error.constructor.name}`);
-		return sendErrorToUser(interaction, errorEmbedUser('I had a small network error when messaging Discord. Please run this command again!'));
-	}
-
-	if (error instanceof DiscordAPIError || error instanceof HTTPError) {
-		if (IgnoredCodes.has(error.code)) return;
-
-		client.emit(Events.Error, error);
-	} else {
-		logger.warn(`${getWarnError(interaction)} (${interaction.user.id}) | ${error.constructor.name}`);
-	}
-
-	// TODO: More detailed handling, I need to analyze
-	logger.error(error);
-	return sendErrorToUser(interaction, errorEmbedUser('Something went wrong'));
-}
